Update gltext.h to remove the black text backgrnd
Added check to make discard alpha pixels and shadow pixels. 

When the glyph is created, it automatically comes with a black shadow, as well as a transparent background. This code uses two checks: the first removes the transparent background, and the second removes the black shadow. Adding in a different colored background, and adding options for whether to have a text shadow, should be trivial from here.