Restore glVertexAttribPointer before drawing
glVertexAttribPointer is set when gltCreateText is called.  If that pointer is cached over multiple frames, and some other drawing code modifies glVertexAttribPointer between creating the text and drawing the text, then the resulting text draw is extremely garbled and incorrect.  

glVertexAttribPointer must be set for the text draw before drawing.  It cannot be stored at creation time